airflow/requirements.txt
apache-airflow[postgres]==2.10.5 \
--constraint https://raw.githubusercontent.com/apache/airflow/constraints-2.10.5/constraints-3.10.txt

apache-airflow-providers-ssh
pandas
scikit-learn
loguru
python-dotenv
openai
openpyxl
pyyaml